[macos] Build gn with Xcode 13.3
The bots were configured to build gn with Xcode 12.2 beta 2 which
has been out for over two years. Bump to a more recent version of
Xcode (this has no impact on the minimum version of the OS the gn
binary can run on, this is controlled by `-mmacosx-version-min`
compiler flag).

Bug: non
Change-Id: If7db191e83057df280eeee39526d82d61e619d41
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/13701
Reviewed-by: Takuto Ikuta <tikuta@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>