Fix sample iOS project to work on macOS Monterey and M1 macs
Only python3 is available on macOS Monterey, so convert the sample
project to use python3 to work on new version of macOS.

Add an additional `target_environment` gn variable to control if
the build is a device or simulator build. This is needed to build
for simulator on an M1 mac (where the cpu is identified as `arm64`).

Bug: none
Change-Id: Ibc9ffff24ac5c1e5991c19d7c34ca2c2a95a757e
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/14140
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>