Improve generation of Xcode project
Use relative paths in PBXFileReference instead of full paths, which
reduces the size of the generated Xcode (from 19Mb to 16Mb for the
Chromium project).

Ensure that the "Products" group is the last item in the main group
and add "productRefGroup" property to the main project object. Both
together allow Xcode to hide it thus reducing visual clutter in the
application.

List the source file directly from the main group object, removing
the "Source" group, and one level of folder in the Xcode UI further
reducing visual clutter (this rely on Xcode hiding the "Products"
group).

Bug: chromium: 1331345
Change-Id: Ia6734af8451e82fe3d01ae0ac4ef0c7ff56dd0d2
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/14141
Reviewed-by: Brett Wilson <brettw@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@google.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>