Add flag to include additional files in the generated Xcode project
The chromium developer have found it useful to add `*.md` files to
Xcode project so that they can be edited from Xcode. Add an option
to gn Xcode project generator to include them (to remove the need
to use a post-processing of the project in Chromium).

Bug: none
Change-Id: I0ac0eb6529735292b592cff4654259a2b1a61ff9
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/14160
Reviewed-by: Takuto Ikuta <tikuta@google.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>