Format lists in assignments with binary operations
The content of lists in assignments to the regular variables
sources or deps should be sorted even when another list or
variable is concatenated to it.

This traverses the content of a BinaryOpNode recursively
to repeat the formatting on both sides of the node when
formatting is applicable.

Bug: 298
Change-Id: Ia659f8024e05a546347a4c30f316048176c8067d
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/14180
Commit-Queue: Brett Wilson <brettw@chromium.org>
Reviewed-by: Brett Wilson <brettw@google.com>
Reviewed-by: Brett Wilson <brettw@chromium.org>