Avoid ACCESS_DENIED when writing build.ninja on Windows
Windows is more fussy about writing to files that are open which cause
updates to build.ninja to reliably fail after recent changes. This fixes
the issue by closing build_ninja_file (build.ninja) before writing to
it.

Minimal tests of WriteFileAtomically were also added in hopes that they
would reveal the failure, but that was before the nature of the bug was
known. Unfortunately this bug would only appear in an integration test.

The real issue was found by modifying the code to spin in a loop when
the failure was hit, attaching a debugger, pausing to inspect variables,
and then using sysinternals' "handle.exe" to find which file was locked.

Bug: https://bugs.chromium.org/p/chromium/issues/detail?id=1351858
Change-Id: I99e7816771de0feb00ec9889640c9cdcaae6f3be
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/14380
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>