Add export compile commands .gn setting.
Adds a "export_compile_commands" setting to the .gn file. This can be
used as a consistent way to generate a compilation database for projects
that require this (the existing --export-compile-commands switch is
intended for end-users to use with editor integration).

The new variable uses label patterns for better consistency with the
rest of the system. The switch value takes precendence if present.

Bug: 111

Change-Id: Ieec41f55e9a140066e8b3e3f69115ab087f52172
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/14520
Commit-Queue: Brett Wilson <brettw@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@google.com>