Add to export_compile_commands from cmdline.
Adds a new switch to append to the export_compile_commands variable in
the dotfile. This is for users who may want to add more targets than
their project defines or, more commonly, for users to define something
when their project defines nothing.

Changes the previous behavior of the legacy filter switch. Previously it
overrode the new pattern behavior. But that prevents rolling out this
feature in a non-breaking way for users who use the switch. The new
behavior constructs the union of all ways the targets can be specified.

In support of this, adds the ability for the CommandLine to allow
multiple values for a given switch. The current behavior where the last
one is returned when querying for a single value is preserved.

Rename GetSwitchValueASCII to GetSwitchValueString and convert to UTF-8
on Windows. GN supports non-ASCII file paths which are either platform
native or UTF-8 on Windows, and this handling matches that rule. This
allows non-ASCII patterns to be specified (though the user is
responsible for getting the correct encoding on e.g. Linux).

Bug: 111, 302

Change-Id: I8ff60260b93d908b2f519c7d83ff8eafcfc5be79
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/14540
Reviewed-by: Takuto Ikuta <tikuta@google.com>
Commit-Queue: Brett Wilson <brettw@chromium.org>