[swift] Consider swift bridge_header as a source file
The target that builds .swift files may use an optional bridge header.
This header is used to import Objective-C types in Swift when modules
are not used (which Chromium currently does not support).

As this header may include any file from the project, it needs to be
considered as a source file (it cannot be listed in sources directly
as mixed languages is not supported for .swift files).

This fixes `gn check` for the target that uses .swift with a bridge
header (which happens in Chromium).

Bug: 305
Change-Id: Iac64c88258302b4f39d8feafc5eb020cfea8efa8
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/14660
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@google.com>