Add TargetPublicPair class.
This is conceptually similar to InheritedLibraries, except that this
separates the temporary UniqueVector<> used during construction, from
the final result, which is a simple vector of (target, is_public) pairs
instead, which is smaller to store and faster to parse during target
resolution.

This will be used in a future CL that moves and caches target resolution
computations to a dedicated class.

Bug: None

Change-Id: I0f7840d105e812a7d43c415ce5e71e43fec1268e
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/14880
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Brett Wilson <brettw@chromium.org>