Fix escaping of interpreter, script, & source dir
The wrong escaping mode is used in a few contexts, which results in
failed builds. This is happening in particular to folks whose sources
are inside a directory with a space in it, and whose builds are not
contained within that directory.

Such builds encode the surrounding filesystem structure into the ninja
files, which is unfortunate but unavoidable in some environments. Fix
the escaping mode in this case.

Bug: 325
Change-Id: Iab53514a3c5cbdb6e7a1d995469e2a75e44924dd
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/15080
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>