[action][data deps] Make data-deps order-only deps of the action outputs
This change makes action() data_deps behave in the same manner as
executable() data_deps:  They are order-only deps of the outputs of the
action() itself, and not order-only deps of the stampfile for the
action.

After this change, when build the output of an action:

  ninja path/to/action/output

the data_deps will also be verified as up-to-date, and built if
necessary.

Before this change, the data_deps for the action() would not have been
verified as up-to-date.

Change-Id: I30e9aff8af60df233294ed40f119889af5ee0ad5
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/15720
Reviewed-by: Takuto Ikuta <tikuta@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>