[action] Add test for data_deps of an action target
Add a new test that shows how data_deps of an action are added as ninja
order-only deps for the rules created for the action.

This illustrates the currently behavior, where the order-only deps are
added to the stamp file rule, not to the action's own rule.

Change-Id: I0d9d7966621a5002e5f9df5b0718db7a8ea78e74
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/15740
Reviewed-by: Takuto Ikuta <tikuta@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>