[docs]: Mention implicit names in style guide
The gn reference explicitly states that part of the preferred style for
adding deps is to omit the colon when possible. However, this is not
mentioned in the style guide where one would usually go searching for
this information.

Mention a preference to omit colons from deps when possible in the deps
portion of the gn style guide.

Bug: none
Change-Id: Ia4fc02ea47ff9a9d21268dcc6a8e977b4cb500d3
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/16020
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>