Add `root_patterns` list to build configuration.
This CL adds a new optional `root_patterns` argument in
the top-level `.gn` file, as well as corresponding
`--root-pattern=PATTERN` command-line arguments.

This can be used to restrict the list of targets that
are added to the build graph when BUILD.gn files are evaluated
in the default toolchain.

The default behavior, is to add all targets defined in
the BUILD.gn files that are evaluated in the default toolchain,
including all their transitive dependencies.

When a non-empty list of patterns is used, these will be used
to filter the targets by their label, in the default toolchain
(it is thus an error to specify a toolchain suffix in these
patterns).

For example, on a complex Fuchsia build plan, using
`gn gen out/default --root-target=//:*` changes `gn gen`
time from 18s to 6s, reducing the generated Ninja file
count from 21907 to 8376 files, which reduces the Ninja
startup time by ... seconds.

Change-Id: Ic727b64e8d899d0dd88113576c1c8ae6e035d326
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/16060
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Takuto Ikuta <tikuta@google.com>