Move InvokePython() function to its own source file.
This makes the function usable from other source files,
as used in a future CL that will add a new type of file
writer.

The new function is placed in the `internal` namespace
to be consistent with `internal::ExecProcess()` which
plays a similar role.

Change-Id: I246122d43cb9bdb88ae3392679716430f8e62f3a
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/16760
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Takuto Ikuta <tikuta@google.com>