[swift] Set `restat = 1` to swift build rules
The swift compiler may leave output file untouched if they are
identical when performing an incremental build. This results
in what should be "no-op" builds to be considered dirty.

Set `restat = 1` on the swift buidl rules to tell ninja that
the compiler perform this optimization and that it should stat
the outputs again and use the result to decide maybe prune the
build graph.

Bug: 362
Change-Id: I97d5096d1541d3f33c7378a7881214f11fece14d
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/16860
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>