[swift] Update `gn check ...` to consider the generated header
The compilation of a Swift module generates an Objective-C header
to allow the interop between Swift and Objective-C. Update checker
to consider this generated header as part of the source_set. This
ensure that missing deps on swift_source_set(...) targets will be
reported by `gn check ...` in Chromium.

  This would prevent regressions caused by missing deps such
  as found and fixed by https://crrev.com/c/5348787.

Fixed: 361
Change-Id: I5394097030e0b08d5ebda58066b16fe2a5bddebf
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/16840
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@google.com>