[ios] Simplify handling of assets catalog
If the source files are listed as part of an assets catalog (i.e. one
element in the path ends with `.xcassets`), then consider the file as
part of an assets catalog.

When generating the Xcode project, if a file in part of an assets
catalog, only list the catalog, not the individual files (Xcode will
discover the files).

Bug: none
Change-Id: Ifaa940e51080ec05bda751a2028c6df2534866ec
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/16900
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>