Support linking frameworks and swiftmodules in Rust targets
Frameworks are straightforward and just need to be added to the rustdeps
line using the RustTool's framework switch. However we note that
weak_frameworks is not somethinng rustc has support for, so we just use
-framework for now to avoid dumping the framework file name into the
commandline and causing more confusing compiler errors.

Swiftmodules require a flag that is passed through to the Rust linker,
but the Rust linker can be a regular linker (link.exe, lld, etc) or can
be a C compiler (clang). If it's the latter, then the C tool's
swiftmodule_switch will work (prefixed with `-Clink-arg=`). However in
the former case, we can not reuse the C tool's switch. So we add
rust_swiftmodule_switch to be specified as a switch that is passed to
the Rust compiler (not its linker).

Bug: 369
Change-Id: I44147172bda2f80ee4f3d89bf32867aaa617adb8
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/17320
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Dana Jansens <danakj@google.com>
Reviewed-by: Takuto Ikuta <tikuta@google.com>