Do not add native dependencies to the library search path
By adding the path to all native dependencies to the library search
path we create collisions if any GN native target has a name that
matches a library specified in `libs`. This creates linking failures
if the wrong library is chosen by the linker.

To add a native library to the search path, for use in #[link], the
`target_out_dir` of the target should be explicitly added to
`lib_dirs`.

Bug: 370
Change-Id: If280641063a5edf3c548e553b350f8bfcc76611c
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/17300
Reviewed-by: Dana Jansens <danakj@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Takuto Ikuta <tikuta@google.com>
Commit-Queue: danakj <danakj@chromium.org>