Markdown optimization:
The GN Reference [1] guide has a lengthy document. When users click on a specific sub-item
to jump to the corresponding position, they need to scroll for a long time to return to the
top of the document. Therefore, an anchor to return to the top has been added to each sub-item.

You can view the changes here [2].

[1] https://gn.googlesource.com/gn/+/main/docs/reference.md
[2] https://github.com/ZhaoSongGOO/MyDocument/blob/main/reference.md#gn-reference
Change-Id: I7a98c95327135ed23b655dd56d9a8e3e80838a6e
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/17420
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Takuto Ikuta <tikuta@google.com>
Reviewed-by: zhao Song <zhaosonggo@gmail.com>