Properly verify runtime_outputs in rust tool definitions.
Ensure that GN will complain when runtime_outputs is set
in a Rust tool type that does not support it. Before this
CL the value was read by the tool, and just ignored (without
any warning about unused variables). This applied to rust_rlib
and rust_staticlib which should not support this argument.

The CL at [1] was reverted, because it broke the Chromium build.
The root cause for this is that it removed runtime_outputs support
from the rust_bin tool by mistake. Adding proper tests here
should prevent similar errors in the future.

[1] https://gn-review.googlesource.com/c/gn/+/17401

+ Improve the FunctionToolchain.RuntimeOutputs test case
  to verify that all C++ and Rust link tools support
  runtime_outputs properly, and that other tools will
  error when it is specified.

Bug: 377
Change-Id: If282807a325bf9ecaa57821a5563b8829687d6c0
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/17540
Reviewed-by: Takuto Ikuta <tikuta@google.com>
Commit-Queue: David Turner <digit@google.com>