Add Target::dependency_output_alias()
In order to support emitting Ninja phony targets instead of stamp,
this CL allows the Target class to provide either a
dependency_output_file() or a dependency_output_alias() as their
output. Only one of these can be set, though both can be empty
under certain circumstances. Code that doesn't care can also use
dependency_output().

This also modifies the behaviour of certain Target methods in
the case where the "no_stamp_files" build settings is enabled,
however the latter is currently disabled (see previous CL),
which means that no change in behaviour should occur due to
this CL.

The next CL will hook the target generator to these methods
and will ensure the phony rule generation works as expected.

For testing the output of `gn gen` has been compared before
and after this CL was applied with the Fuchsia source tree
to verify that the output is the same.

Bug: 172
Change-Id: Ide890dd4d016e7e27e633dc67906a87aa8d1acd4
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/12865
Reviewed-by: Philipp Wollermann <philwo@google.com>
Commit-Queue: Takuto Ikuta <tikuta@google.com>
Reviewed-by: Takuto Ikuta <tikuta@google.com>